SerialPort: write() returns size_t
The values sizeof(unsigned int) and sizeof(size_t) might differ on some
platforms a lot. The ssize_t (the maximal possible return value from
write()) is just 1-bit shortert than size_t. Thus, when using unsigned
int, we have a chance to overflow (however, practically not very possible).

Fixes: 9246d6af4ab6 ("SerialPort: new class")

+++ src/io/SerialPort.cpp
@@ -202,7 +202,7 @@ void SerialPort::open()
 	}
 }
 
-unsigned int SerialPort::write(const char *buffer, size_t size)
+size_t SerialPort::write(const char *buffer, size_t size)
 {
 	ssize_t ret = ::write(m_fd, buffer, size);
 
@@ -219,7 +219,7 @@ unsigned int SerialPort::write(const char *buffer, size_t size)
 	return ret;
 }
 
-unsigned int SerialPort::write(const string &data)
+size_t SerialPort::write(const string &data)
 {
 	return write(data.c_str(), data.length());
 }
